+++ tests/testthat/test-e1071.R
@@ -22,3 +22,17 @@ test_that("svm", {
   expect_equal(predict(twidlr_fit, data = d),
                e1071:::predict.svm(origin_fit, newdata = d))
 })
+
+
+test_that("naiveBayes", {
+  
+  d <- datasets::mtcars
+  d$cyl_gr_6 <- as.factor( ifelse(d$cyl > 6, 1, 0) )
+  
+  twidlr_fit <- twidlr::naiveBayes(d, cyl_gr_6 ~ .)
+  origin_fit <- e1071::naiveBayes(cyl_gr_6 ~ ., d)
+  
+  expect_equal(twidlr_fit$tables, origin_fit$tables)
+  expect_error(predict(twidlr_fit))
+  expect_equal(predict(twidlr_fit, data = d), e1071:::predict.naiveBayes(origin_fit, newdata = d))
+})