+++ NEWS.md
@@ -10,3 +10,4 @@
     * `lme4::glmer` (thanks to [Mathew Ling](https://github.com/Lingtax/))
     * `quantreg::rq, nlrq, rqss, crq` (thanks to [Joran Elias](https://github.com/joranE))
     * `e1071::svm`
+    * `e1071::naiveBayes` (thanks to [Bjoern Oettinghaus](https://github.com/erblast))